Restructure product unit tests folder
