+++ src/test/java/com/commercetools/sync/sdk2/products/AssetUtils.java
@@ -1,4 +1,4 @@
-package com.commercetools.sync.sdk2.products.utils;
+package com.commercetools.sync.sdk2.products;
 
 import static com.commercetools.sync.sdk2.products.utils.CustomFieldsUtils.createCustomFieldsDraft;
