+++ src/test/java/com/commercetools/sync/sdk2/products/utils/productvariantupdateactionutils/attributes/AttributeFixtures.java
@@ -1,6 +1,6 @@
 package com.commercetools.sync.sdk2.products.utils.productvariantupdateactionutils.attributes;
 
-import static com.commercetools.sync.sdk2.products.utils.ProductSyncMockUtils.createObjectFromResource;
+import static com.commercetools.sync.sdk2.products.ProductSyncMockUtils.createObjectFromResource;
 
 import com.commercetools.api.models.product.Attribute;
